Disables vanguard when exit league completely.
Now, the script will disable the SVG service on PC startup, so it will be inactive every time you boot your computer. To play, you'll need to start the game from the script to activate the services and keep it running. This way, when you exit the game, the script will handle turning off the services again."